Duck type TranslatorAwareInterface
Starting with zend-mvc 3, i18n features are considered opt-in and optional.
However, several official helpers in zend-view are translator-aware, and
optionally expose translation features if a translator is composed.

This patch does the following:

- Introduces `TranslatorAwareTrait`, which encapsulates the logic for composing
  a translator and translation text domain. This logic was duplicated across
  multiple helpers already, and is simply extracted from it.
- Updates `FlashMessenger`, `HeadTitle`, and the navigation `AbstractHelper` to:
  - Remove the explicit implementation of `TranslatorAwareInterface`.
  - Use the new `TranslatorAwareTrait`, effectively duck typing
    `TranslatorAwareInterface`.
- Updates `HelperPluginManager` to inject a translator for helpers that either
  explicitly implement `TranslatorAwareInterface` or duck type it.
