+++ src/Helper/FlashMessenger.php
@@ -1,22 +1,23 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @link      http://github.com/zendframework/zend-view for the canonical source repository
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
 namespace Zend\View\Helper;
 
 use Zend\Mvc\Controller\Plugin\FlashMessenger as PluginFlashMessenger;
-use Zend\I18n\View\Helper\AbstractTranslatorHelper;
 
 /**
  * Helper to proxy the plugin flash messenger
+ *
+ * Duck-types against Zend\I18n\Translator\TranslatorAwareInterface.
  */
-class FlashMessenger extends AbstractTranslatorHelper
+class FlashMessenger extends AbstractHelper
 {
+    use TranslatorAwareTrait;
+
     /**
      * Default attributes for the open format tag
      *
@@ -149,9 +150,9 @@ class FlashMessenger extends AbstractTranslatorHelper
         }
 
         // Flatten message array
-        $escapeHtml      = $this->getEscapeHtmlHelper();
-        $messagesToPrint = [];
-        $translator = $this->getTranslator();
+        $escapeHtml           = $this->getEscapeHtmlHelper();
+        $messagesToPrint      = [];
+        $translator           = $this->getTranslator();
         $translatorTextDomain = $this->getTranslatorTextDomain();
         array_walk_recursive(
             $messages,
